+++ custom-recipes/image-conversion/recipe.py
@@ -2,11 +2,12 @@ from dataiku.customrecipe import get_recipe_config
 from PIL import Image
 from io import BytesIO
 import logging
-from ocr_recipes_io_utils import get_input_output
-from ocr_utils import convert_image_to_greyscale_bytes
-from ocr_utils import image_conversion_parameters
-from ocr_utils import pdf_to_pil_images_iterator
-from ocr_constants import Constants
+from text_extraction_ocr_utils.recipes_io_utils import get_input_output
+from text_extraction_ocr_utils import convert_image_to_greyscale_bytes
+from text_extraction_ocr_utils import image_conversion_parameters
+from text_extraction_ocr_utils import pdf_to_pil_images_iterator
+from text_extraction_ocr_utils import Constants
+
 
 logger = logging.getLogger(__name__)