+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## [Version 2.1.0](https://github.com/dataiku/dss-plugin-tesseract-ocr/releases/tag/v2.1.0) - Feature release - 2023-06
 
-- Add a recipe to extract text directly (not OCR) from multiple types of files
+- Add a recipe to extract text content from various file types
 
 ## [Version 2.0.0](https://github.com/dataiku/dss-plugin-tesseract-ocr/releases/tag/v2.0.0) - Feature release - 2023-06